Extra_Long_Factorials_Statement
Problem Statement:
The factorial of the integer n, written n! , is defined as:
   n! = n * (n-1) * (n-2) * (n-3) * ............ * 3 * 2 * 1 
Calculate and print the factorial of a given integer.

For example, if n= 30, we calculate 30 * 29 * 28 .......... * 3 * 2 * 1 and get 265252859812191058636308480000000

Function Description:
Complete the extraLongFactorials function in the editor below. It should print the result and return.
extraLongFactorials has the following parameter(s):
    n: an integer


https://www.hackerrank.com/challenges/extra-long-factorials/problem